Break out all sensors to their own file
